cadvisor: build: fix broken build with Makefile
When using `make build` to compile the source,
build fails with error/s. The reason is because
of GO_CMD variable is set to 'install' as default,
which is not valid when running `make build` or
`make`.

Depending on the valid git tag, release.sh will
set the GO_CMD variable. If valid git tag is
available e.g.: vx.y.z.beta1, then GO_CMD is
set to `build`.

If valid git tag is not available, release.sh
will any way bails out serving the purpose.
So drop using GO_CMD, to fix the default build.